Crate passivetotal

This is the Rust implementation of a client for the PassiveTotal API. Usage is provided through passivetotal::client::PTClient

Before running anything, you will need to create a JSON configuration. See the config module documentations for details.

Please see the passivetotal::client::PTClient documentation for the available methods.

See the response module documentation for the specification for the response structs.

See the API documentation for more resources.

Examples

use passivetotal::config::read_config;
use passivetotal::client::PTClient;

let conf = match read_config() {
    Ok(conf) => conf,
    _ => panic!("Please create your config at ~/.config/api_config.json"),
}
let client = PTClient::new(conf);
let response = match client.get_pdns("passivetotal.org") {
    Ok(response) => response,
    _ => panic!("Something bad happened in the JSON response"),
};

Modules

client
The client module exposes the PTClient struct which is the main interface to use, and contains all methods necessary to make calls to the API.

config
It is necessary to create the correct configuration before using the API.

constants
Warning: incomplete

response
The response module contains all the structs that will be returned as a response for all API calls.