Module passivetotal::client

The client module exposes the PTClient struct which is the main interface to use, and contains all methods necessary to make calls to the API.

Please see the passivetotal::client::PTClient documentation

Structs

PTClient
This is the client exposed to the user for abstracting the passivetotal API.

Enums

ResponseError