Module passivetotal::response

The response module contains all the structs that will be returned as a response for all API calls.

These were created programmatically through a tool json2rust, which takes a JSON document and builds the corresponding Rust struct. This simplified the process greatly, since a lot of these documents are extensive.

This is the heart of the client, and maps all responses to the sort of data they contain. The fields are in camelCase to correspond with the fields returned via the API.

Structs

AccountResponse

ActionClassificationResponse

ActionDDNSResponse

ActionEverCompromisedResponse

ActionMonitorResponse

ActionSinkholeResponse

ActionTagResponse

HostAttributeComponentResponse

HostAttributeComponentResult

HostAttributeTrackerResponse

HostAttributeTrackerResult

MalwareResponse

MalwareResult

OSINTResponse

OSINTResult

PDNSResponse

PDNSResult

PDNSUniqueResponse

Pager

Registrant

SSLCertHistoryResponse

SSLCertHistoryResult

SSLCertResponse

SubdomainsResponse

WhoisResponse